// src/app/api/models/dynamic-theme-variant-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * DTO for theme variant information
 */
export interface DynamicThemeVariantDto {

  /**
   * Variant ID
   */
  id?: string;

  /**
   * Whether this is the default variant for the theme
   */
  isDefault?: boolean;

  /**
   * Variant name (Default, Blue, Green, Purple, Red)
   */
  name?: string;

  /**
   * Dynamic tokens for this variant (overrides base theme tokens)
   */
  tokens?: {
[key: string]: string;
};
}