// src/app/api/models/dynamic-user-theme-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DynamicThemeDto } from '../models/dynamic-theme-dto';
import { DynamicThemeVariantDto } from '../models/dynamic-theme-variant-dto';

/**
 * DTO for user theme preferences
 */
export interface DynamicUserThemeDto {

  /**
   * Selected dark theme information
   */
  darkTheme?: DynamicThemeDto | null;

  /**
   * Selected Dark Theme ID
   */
  darkThemeId?: string | null;

  /**
   * Selected dark variant information
   */
  darkVariant?: DynamicThemeVariantDto | null;

  /**
   * Selected Dark Theme Variant ID
   */
  darkVariantId?: string | null;

  /**
   * Selected light theme information
   */
  lightTheme?: DynamicThemeDto | null;

  /**
   * Selected Light Theme ID
   */
  lightThemeId?: string | null;

  /**
   * Selected light variant information
   */
  lightVariant?: DynamicThemeVariantDto | null;

  /**
   * Selected Light Theme Variant ID
   */
  lightVariantId?: string | null;

  /**
   * Whether to sync with system settings
   */
  syncWithSystem?: boolean;

  /**
   * User ID
   */
  userId?: string;
}